Refactor: Move module visitors to `visitor` module
