+++ src/asm/generate/class.rs
@@ -1,8 +1,8 @@
-use crate::asm::generate::byte_vec::{ByteVec, ByteVecImpl};
+use crate::asm::generate::bytes::{ByteVec, ByteVecImpl};
 use crate::asm::generate::field::FieldWriter;
 use crate::asm::generate::method::MethodWriter;
 use crate::asm::generate::symbol::SymbolTable;
-use crate::asm::generate::ByteVecGen;
+use crate::asm::generate::bytes::ByteVecGen;
 use crate::asm::node::access_flag::{
     AccessFlags, ClassAccessFlag, FieldAccessFlag, MethodAccessFlag,
 };
