+++ src/asm/visitor/mod.rs
@@ -2,3 +2,4 @@ pub mod class;
 pub mod field;
 pub mod method;
 pub mod signature;
+pub mod module;
